Resume Analyzer and Job Recommendation System/requirements.txt
flask==3.0.0
PyPDF2==3.0.1
python-docx==1.1.0
sentence-transformers==2.2.2
nltk==3.8.1
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
Werkzeug==3.0.1
